add NSPopover
